tests/wpt/web-platform-tests/mathml/relations/css-styling/padding-border-margin/border-001.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>border</title>
<link rel="help" href="https://mathml-refresh.github.io/mathml-core/#layout-algorithms">
<meta name="assert" content="Verify that border is taken into account.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="helper.js"></script>
<script>
  var epsilon = 1;

  setup({ explicit_done: true });
  window.addEventListener("load", runTests);

  function runTests() {
      test(function() {
          var s = measureSpaceAround("mrow-border")
          assert_approx_equals(s.left, 20, epsilon, "left border");
          assert_approx_equals(s.right, 30, epsilon, "right border");
          assert_approx_equals(s.top, 40, epsilon, "top border");
          assert_approx_equals(s.bottom, 50, epsilon, "bottom border");
      }, "Border properties on mrow");

      test(function() {
          var s = measureSpaceAround("mrow-border-shorthand")
          assert_approx_equals(s.left, 20, epsilon, "left border");
          assert_approx_equals(s.right, 20, epsilon, "right border");
          assert_approx_equals(s.top, 20, epsilon, "top border");
          assert_approx_equals(s.bottom, 20, epsilon, "bottom border");
      }, "Border properties on mrow (shorthand)");

      test(function() {
          var s = measureSpaceAround("mrow-border-logical")
          assert_approx_equals(s.left, 20, epsilon, "left border");
          assert_approx_equals(s.right, 30, epsilon, "right border");
          assert_approx_equals(s.top, 40, epsilon, "top border");
          assert_approx_equals(s.bottom, 50, epsilon, "bottom border");
      }, "Border properties on mrow (logical)");

      test(function() {
          var s = measureSpaceAround("mrow-border-logical-shorthand")
          assert_approx_equals(s.left, 20, epsilon, "left border");
          assert_approx_equals(s.right, 20, epsilon, "right border");
          assert_approx_equals(s.top, 30, epsilon, "top border");
          assert_approx_equals(s.bottom, 30, epsilon, "bottom border");
      }, "Border properties on mrow (logical, shorthand)");

    done();
  }
</script>
</head>
<body>
  <div id="log"></div>
  <p>
    <math>
      <mrow>
        <mrow id="mrow-border"
              style="border-left: 20px;
                     border-right: 30px;
                     border-top: 40px;
                     border-bottom: 50px;">
          <mspace width="50px" height="50px"></mspace>
        </mrow>
      </mrow>
    </math>
  </p>
  <p>
    <math>
      <mrow>
        <mrow id="mrow-border-shorthand"
              style="border: 20px;">
          <mspace width="50px" height="50px"></mspace>
        </mrow>
      </mrow>
    </math>
  </p>
  <p>
    <math>
      <mrow>
        <mrow id="mrow-border-logical"
              style="border-inline-start: 20px;
                     border-inline-end: 30px;
                     border-block-start: 40px;
                     border-block-end: 50px;">
          <mspace width="50px" height="50px"></mspace>
        </mrow>
      </mrow>
    </math>
  </p>
  <p>
    <math>
      <mrow>
        <mrow id="mrow-border-logical-shorthand"
              style="border-inline: 20px;
                     border-block: 30px;">
          <mspace width="50px" height="50px"></mspace>
        </mrow>
      </mrow>
    </math>
  </p>
</body>
</html>
